+++ flexiznam/schema/microscopy_data.py
@@ -102,6 +102,7 @@ class MicroscopyData(Dataset):
         project=None,
         project_id=None,
         origin_id=None,
+        id=None,
         flexilims_session=None,
     ):
         """Create a Microscopy dataset
@@ -119,6 +120,7 @@ class MicroscopyData(Dataset):
             project_id: hexadecimal code for the project. Must be in config, can be
                         guessed from project
             origin_id: hexadecimal code for the origin on flexilims.
+            id: hexadecimal code for the dataset on flexilims.
             flexilims_session: authentication session to connect to flexilims
 
         Expected extra_attributes:
@@ -134,6 +136,7 @@ class MicroscopyData(Dataset):
             project=project,
             project_id=project_id,
             origin_id=origin_id,
+            id=id,
             flexilims_session=flexilims_session,
         )